Replace all issuefinder code with new error setup
NOTE: Does not compile!

Changes the issuefinder at its core to store errors directly on the
schema.* types rather than aggregating a list of errors with various
lookups.  This breaks a great deal of the app due to how many places
were tied to the issuefinder's approach to mapping errors.