apperr.Error: Add Propagate method
This will let us know when an error should show flag the parent (an
issue's title or a file's issue) as having an error.  This should make
error reporting to end users more sane, as some "errors", like an issue
being too new, are really not anything concerning, but should still be
treated as issue-level errors for queueing.