queue-batches: Don't iterate issue list
Only attempts to put batches into the queue once instead of iterating
multiple times to try and jam as many issues as we can.  This could
result in slightly smaller batches, but I'm seeing an immediate bug here
that would put dupes into the original list, and it's really not worth
fixing just to get a 3950-page batch up to 3990 pages or something.