queue-batches: fix precision issues in testing
The point of the "stale days" computation was to ensure things held too
long get batched, but the test needed a bit more precision than we
really care about.  It only causes problems when twenty days ago...
includes a time change.  Because then it's 20 days ago, but *not* 20*24
hours ago.  ARGH.