cmd/run-jobs: add in-house scan folder watcher
This change will automate the move of (likely) valid scanned issues into
NCA so that they don't have to be addressed manually each time a batch
of scans is completed.  The scans directory is stricter about what may
be present than the "sftp" dir, so is meant for items that have already
had some amount of quality control and include a TIFF, not just PDFs.
i.e., this won't change how publisher-uploaded PDFs are treated.