Make models.Issue.Save require action data
Since we usually want action data, this makes callers have to either put
in extra data and use Save(), or explicitly say SaveWithoutAction(),
making it clear the change is not going to show up in the actions log.

Adds action logging for issue claim and unclaim as well as system
changes to workflow step.