Add an option to re-batch issues
Adds a new workflow step for issues to indicate they need to be
rebatched, and adds a flag to `queue-batches` to only operate on issues
in this workflow step.

This is effectively a separate "bucket" for issues so that when creating
new batches we can batch only issues which have just been flagged as
"redo" issues, rather than batching everything that's ready. This should
make it easier to validate the "rebatched" issues.