jobs: don't log or exit when prepping jobs
Takes the status/state checks out of the two "prepare*JobAdvanced"
functions. First, these using `logger.Fatalf` was a bad idea, as that
will just terminate the application (I think I meant this to be a
placeholder, not a permanent solution). Second, it doesn't make sense to
check status and fail anyway - this means needing to do a lot more
active monitoring. Instead we'll refactor so jobs are forced through
functions that set state as needed to begin with.