Set jobs to pending based on pipeline status
The first job in a pipeline starts pending. Then whenever a job
finishes, it checks the pipeline's other jobs: if any are still
unfinished, nothing happens; if there are no more jobs in the pipeline,
the pipeline itself is updated to be completed; and if there are any
jobs in the pipeline with the finished job's sequence + 1, they're set
to pending.