models: rename x.Job() to x.BuildJob()
Replaces issue, batch, and job's Job() function with BuildJob() so it's
clearer what's going on - these all generate an empty, unsaved Job. The
problem I'm solving here is that there is also an x.Jobs() method on
some models that *fetches* jobs from the database, and x.Job() is just
too close in name for how different it is in usage.