jobs: add "PRFatal" to signal immediate failure
Sets up a new process response to let jobs signal a fatal error that
can't be overcome by a retry.

This is used very spartingly: *only* when it's exceedingly unlikely a
temporary problem might exist, like a file existing that shouldn't or
database state that wasn't set up properly for the job. Most database
errors and disk I/O problems, on the other hand, warrant a retry.