upload: adjust logging on unexpected jobs
When queueing up an issue, if any job exists that isn't
JobStatusFailedDone, we throw a critical error. But a slow filesystem,
or a lot of pending jobs, can mean pending jobs sit a while - long
enough for the queue process to look for them again and try to queue
them up a second time. This must obviously be avoided, but it's not a
critical problem.