Add a mutex around uploaded issue queueing
This is a bit of a sledgehammer approach, but it will work: the uploaded
issue handler can occasionally receive two POSTs for the same issue. I'm
not sure how end-users are getting around the double-submit protection
in JavaScript, but somebody has been basically smashing the "Queue
Issue" button insanely fast, possibly before the page is even done
loading, so we're stuck preventing the problem on the server side.