Rewrite agent "test" to be useful
Adds some utility to the tester so we can continue to validate the ONI
Agent endpoint, but also run all the commands NCA is going to be
running. This is still probably not a long-term project, but it'll help
test as well as checking things like ONI Agent logs for jobs that fail.