test: replace batch full names first!
Fixes a problem where the map order might first change the shortened
name, and then be unable to find the longer name.

e.g., if the batch is "batch_oru_20241205SomeLongWeirdName_ver01", the
short replace will rewrite "SomeLongWeirdName" to, say,
"Pages8TitlesBlah", resulting in
"batch_oru_20241205Pages8TitlesBlah_ver01". But then the long replace is
searching for the full name which no longer exists, leaving things like
the date in a bad state.