titlehandler: fix MARC upload reading
Uses a bytes.Reader instead of a bytes.Buffer, initializing it right
from the uploaded data so there's no need to write, and it's immediately
ready to read.

Also fixes the hilarious mistake of trying to parse the uploaded file
instead of the buffer. The uploaded file is already read, and its
position is at the end of the file. OOPS.