jobs: handle ONI Agent "no job needed" response
We never handled the agent's potential to return job id of -1 when a job
isn't needed, and so isn't sent to ONI. This fixes the job which was
waiting on ONI Agent so that if the job ID was -1, it just calls the
"wait" a success.

Funny story here. This was only uncovered because of a different bug
(still tracking it down) that queues up jobs incorrectly. So... a bug
was found because of sloppy code. Makes you think.