derivatives/jp2: don't provide default bin paths!
This fix makes configuring the JP2 derivatives a bit more consistent,
and makes it easier to see when a binary path isn't being set properly.
With defaults set up, it would be (and was) easy for hard-coded settings
to mask potential problems on different platforms where the paths aren't
going to be so simple.